-Added integer cast to confusion matrix

+++ VisionTools.py
@@ -675,7 +675,7 @@ def confusion_matrix(y_label, y_predict, label_for_print=None):
     :return: Confusion matrix as nparray
     :rtype: np.core.multiarray.ndarray
     """
-    y_label, y_predict = np.array(y_label), np.array(y_predict)
+    y_label, y_predict = np.array(y_label).astype(np.int), np.array(y_predict).astype(np.int)
     n_max = max(y_label.max(), y_predict.max()) + 1
     cm = np.zeros((n_max, n_max), dtype=np.int)
     np.add.at(cm, [y_label, y_predict], 1)
